Moved costumes adapters to a new file costumes.js
